complete convo cheched in

+++ base/views.py
@@ -1,15 +1,21 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
-from .models import Room, Topic
 from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic
 from .forms import RoomForm
-from django.contrib.auth import authenticate, login
+
 
 # Create your views here.
 
 def loginPage(request):
-
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home') 
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -17,11 +23,42 @@ def loginPage(request):
         try:
             user = User.objects.get(username=username)
         except:
-            messages.error(request, 'user does not exist')
-         
+            messages.error(request, 'user does not exist.')
 
-    context = {}
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'username OR password does not exits')
+        
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm(request.POST)
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Try Again! incorrect pass')
+      
+
+    return render(request, 'base/login_register.html', {'form': form})
+                 
+
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -41,9 +78,11 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    room_messages = room.message_set.all()
+    context = {'room': room, 'room_messages': room_messages}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form  = RoomForm()
     if request.method == 'POST':
@@ -54,9 +93,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form  = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('you are not allowed user')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -67,8 +110,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('you are not allowed user')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
